requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.4
django-phonenumber-field==8.0.0
frozenlist==1.5.0
idna==3.10
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.2
multidict==6.1.0
phonenumbers==8.13.51
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytube==15.0.0
redis==5.2.1
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
twilio==9.4.0
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
yt-dlp==2024.12.13
